Выполнила ДЗ4 Модуляция. Оптимизировала функцию получения массива случайных значений случайной длины.
